windows-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
aiosqlite==0.20.0
cryptography==43.0.3
customtkinter==5.2.2
Pillow==11.0.0
pystray==0.19.5
aiofiles==24.1.0
pydantic==2.9.2
